fix(ids): Externalizes NullUUID.
