+++ id/null_id.go
@@ -1,0 +1,40 @@
+package id
+
+import (
+	"database/sql/driver"
+	uuid "github.com/satori/go.uuid"
+)
+
+// NullUUID can be used with the standard sql package to represent a
+// UUID value that can be NULL in the database
+type NullUUID struct {
+	UUID  uuid.UUID
+	Valid bool
+}
+
+// Scan implements the sql.Scanner interface.
+func (u *NullUUID) Scan(src interface{}) error {
+	if src == nil {
+		u.UUID, u.Valid = uuid.Nil, false
+		return nil
+	}
+
+	// Delegate to UUID Scan function
+	u.Valid = true
+
+	switch src := src.(type) {
+	case uuid.UUID:
+		return u.UUID.Scan(src.Bytes())
+	}
+
+	return u.UUID.Scan(src)
+}
+
+// Value implements the driver.Valuer interface.
+func (u NullUUID) Value() (driver.Value, error) {
+	if !u.Valid {
+		return nil, nil
+	}
+	// Delegate to UUID Value function
+	return u.UUID.Value()
+}
